% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_test_across_groups.R
\name{enrichment_test_across_groups}
\alias{enrichment_test_across_groups}
\title{Test a list of protein identifiers for enrichment in multiple groups}
\usage{
enrichment_test_across_groups(
  protein_foreground_list = NULL,
  factor_minimum_explained_variance = 0,
  test_across = c("sex", "ancestry"),
  protein_background = NULL,
  n_cores = 1
)
}
\arguments{
\item{protein_foreground_list}{Named list. Different protein foregrounds to be tested for enrichment of UK Biobank study characteristics sequentially. The name of each list item is used as a grouping variable. See examples for details.}

\item{factor_minimum_explained_variance}{Numeric. Defaults to \code{0}. Value between 0-1 to require a minimum of explained variance by certain factors to be considered.}

\item{test_across}{String. Defaults to "sex". Whether to test for enrichment stratified by sex or genetic ancestry. Either "sex" or "ancestry" is accepted.}

\item{protein_background}{Vector of strings. Optional list of proteins to be used as background.}

\item{n_cores}{Integer. Defaults to \code{1}. Number of cores to use to parallelize enrichment testing. Passed to \code{mc.cores} of \code{parallel::mclapply}.}
}
\value{
A data.table containing enrichment results for all groups in input data.
}
\description{
This function helps calculating protein characteristics enrichment for multiple groups. The output of this function can be plotted using \code{plot_enrichment_results()}.
}
\examples{
\dontrun{

protein_foreground_list <- list(
  group1 = c("lep", "tmprss15", "fam3b"),
  group2 = c("pcsk9", "apoa4", "lep", "tmprss15", "fam3b")
)

enrichment_test_across_experiments(
  protein_foreground_list
)
}
}
