% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fasting_study_results}
\alias{fasting_study_results}
\title{Fasting Study Example Data}
\format{
\subsection{\code{fasting_study_results}}{

A data frame with X rows and Y columns
}
}
\source{
<TODO_ADD_LINK>
}
\usage{
fasting_study_results
}
\description{
Plasma proteins that show significant difference in abundance compared to
baseline after prolonged fasting. Day 10 is after recovery from fasting.
}
\keyword{datasets}
