% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{participant_characteristics_labels}
\alias{participant_characteristics_labels}
\title{UK Biobank Study Participant Characteristics}
\format{
\subsection{\code{participant_characteristics_labels}}{

A data frame with X rows and Y columns
}
}
\source{
<TODO_ADD_LINK>
}
\usage{
participant_characteristics_labels
}
\description{
A list of characteristics in UK Biobank that were used as the basis for the
the analysis of factors influencing plasma protein levels.
}
\keyword{datasets}
