% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_enrichment_results.R
\name{plot_enrichment_results}
\alias{plot_enrichment_results}
\title{Plot enrichment results for multiple groups}
\usage{
plot_enrichment_results(
  enrichment_results = NULL,
  pvalue_adjustment = "bonferroni"
)
}
\arguments{
\item{enrichment_results}{data.table. Output from \code{enrichment_test_across_groups()}, restricted to a single population (of either sex or genetic ancestry).}

\item{pvalue_adjustment}{string. Defaults to "bonferroni". Passed to \code{p.adjust()}. See that function's help page for options.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot enrichment of UK Biobank study characteristics across several groups.
}
