% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prete-package.R
\docType{package}
\name{prete-package}
\alias{prete}
\alias{prete-package}
\title{prete: Test for enrichment of UK Biobank Study Characteristics in Protein Lists}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Carl Beuchel \email{carl.beuchel@charite.de} (\href{https://orcid.org/0000-0003-3224-3894}{ORCID})

Authors:
\itemize{
  \item Maik Pietzner \email{mark.pietzner@bih-charite.de} (\href{https://orcid.org/0000-0003-3437-9963}{ORCID})
}

}
\keyword{internal}
