% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{protein_mapping_table}
\alias{protein_mapping_table}
\title{Olink Protein Mapping Table}
\format{
\subsection{\code{protein_mapping_table}}{

A data frame with X rows and Y columns:
\describe{
\item{col}{col meanind}
\item{col2, col3}{,multiple col meanings}
}
}
}
\source{
<TODO_ADD_LINK>
}
\usage{
protein_mapping_table
}
\description{
A mapping table containing protein identifiers that can be used to check own
data against the background used to run the enrichment analysis.
}
\keyword{datasets}
