% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{variance_decomposition_background}
\alias{variance_decomposition_background}
\title{Protein Determinants Data}
\format{
\subsection{\code{variance_decomposition_background}}{

A data frame with X rows and Y columns
}
}
\source{
<TODO_ADD_LINK>
}
\usage{
variance_decomposition_background
}
\description{
Explained variance in protein levels by various participant's characteristics.
}
\keyword{datasets}
